fixed create specific for windows and linux
